Restruct module namespace and cleanup file dependencies
