+++ lib/tumugi/helper.rb
@@ -11,10 +11,5 @@ module Tumugi
         [obj]
       end
     end
-
-    def target(type, *args)
-      klass = Tumugi::Plugin.lookup_target(type)
-      klass.new(*args)
-    end
   end
 end
