+++ lib/tumugi/logger.rb
@@ -1,3 +1,4 @@
+require 'logger'
 require 'forwardable'
 
 module Tumugi
