+++ lib/tumugi/mixin/task_mixin.rb
@@ -1,0 +1,10 @@
+require 'tumugi/plugin'
+
+module Tumugi
+  module TaskMixin
+    def target(type, *args)
+      klass = Tumugi::Plugin.lookup_target(type)
+      klass.new(*args)
+    end
+  end
+end
