+++ test/cli_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require_relative './test_helper'
 
 class Tumugi::CLITest < Test::Unit::TestCase
   examples = {
