+++ test/command/run_test.rb
@@ -1,4 +1,5 @@
-require 'test_helper'
+require_relative '../test_helper'
+require 'tumugi/dag'
 require 'tumugi/task'
 require 'tumugi/command/run'
