+++ test/dag_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require_relative './test_helper'
 require 'tumugi/dag'
 require 'tumugi/task'
