+++ test/task_definition_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require_relative './test_helper'
 require 'tumugi/task_definition'
 
 class Tumugi::TaskDefinitionTest < Test::Unit::TestCase
