+++ test/task_test.rb
@@ -1,4 +1,4 @@
-require 'test_helper'
+require_relative './test_helper'
 require 'tumugi/task'
 require 'tumugi/target'
