+++ test/test_helper.rb
@@ -6,6 +6,9 @@ $LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
 require 'test/unit'
 require 'test/unit/rr'
 
+require 'tumugi'
+require 'tumugi/plugin'
+
 def capture_stdout
   out = StringIO.new
   $stdout = out
