+++ tumugi.gemspec
@@ -18,11 +18,10 @@ Gem::Specification.new do |spec|
   spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
   spec.require_paths = ['lib']
 
-  spec.add_runtime_dependency 'activesupport', '~> 4.2'
-  spec.add_runtime_dependency 'ruby-graphviz', '~> 1.2.2'
-  spec.add_runtime_dependency 'thor', '~> 0.19.1'
   spec.add_runtime_dependency 'parallel', '~> 1.8.0'
   spec.add_runtime_dependency 'retriable', '~> 2.1'
+  spec.add_runtime_dependency 'ruby-graphviz', '~> 1.2.2'
+  spec.add_runtime_dependency 'thor', '~> 0.19.1'
 
   spec.add_development_dependency 'bundler', '~> 1.11'
   spec.add_development_dependency 'rake', '~> 10.0'
